Move a bunch of stuff to TensorOperations
