+++ src/tensors/Tensor.java
@@ -1,8 +1,6 @@
 package tensors;
 
 import functions.GeneralFunction;
-import functions.commutative.Product;
-import functions.commutative.Sum;
 import functions.endpoint.Variable;
 import org.jetbrains.annotations.NotNull;
 import output.OutputFunction;
@@ -15,7 +13,7 @@ import java.util.*;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
-import static tools.MiscTools.minimalSimplify;
+import static tensors.TensorOperations.*;
 
 public class Tensor extends GeneralFunction {
 
@@ -140,36 +138,6 @@ public class Tensor extends GeneralFunction {
 		}
 	}
 
-	public static Tensor modifyWith(Tensor seed,
-									Function<String, String> indexModifier,
-									Function<Boolean, Boolean> contravarianceModifier,
-									Function<Integer, Integer> rankModifier,
-									Function<Tensor, GeneralFunction> tensorElementModifier,
-									Function<GeneralFunction, GeneralFunction> functionElementModifier,
-									boolean simplifyFunctions) { // TODO move the if statement up a level for efficiency
-		if (seed.rank > 1)
-			return new Tensor(
-					indexModifier.apply(seed.index),
-					contravarianceModifier.apply(seed.isContra),
-					rankModifier.apply(seed.rank),
-					Arrays.stream(seed.elements)
-							.map(function -> tensorElementModifier.apply((Tensor) function))
-							.toArray(GeneralFunction[]::new)
-			);
-		else
-			return new Tensor(
-					indexModifier.apply(seed.index),
-					contravarianceModifier.apply(seed.isContra),
-					rankModifier.apply(seed.rank),
-					Arrays.stream(seed.elements)
-							.map(function ->
-									simplifyFunctions
-									? minimalSimplify(functionElementModifier.apply(function))
-									: functionElementModifier.apply(function))
-							.toArray(GeneralFunction[]::new)
-			);
-	}
-
 
 	private GeneralFunction getElementHelper(LinkedList<Integer> indices) {
 		int idx = indices.removeFirst();
@@ -253,67 +221,9 @@ public class Tensor extends GeneralFunction {
 
 
 	public Tensor scale(GeneralFunction scalar) {
-		return scale(scalar, this);
+		return TensorOperations.scale(scalar, this);
 	}
 
-	public static Tensor scale(GeneralFunction scalar, Tensor tensor) {
-		return modifyWith(tensor,
-				i -> i,
-				i -> i,
-				i -> i,
-				i -> i.scale(scalar),
-				i -> new Product(i, scalar),
-				true);
-	}
-
-	public static Tensor tensorProduct(Tensor... tensors) {
-		Tensor tensor = tensors[0];
-		for (int i = 1; i < tensors.length; i++)
-			tensor = tensorProduct(tensor, tensors[i]);
-		return tensor;
-	}
-
-	private static Tensor tensorProduct(Tensor first, Tensor second) { // TODO check order of resulting indices
-		return modifyWith(second,
-				i -> i,
-				i -> i,
-				i -> i + first.rank,
-				i -> Tensor.tensorProduct(first, i),
-				first::scale,
-				false);
-	}
-
-
-	public static Tensor sum(Tensor... tensors) {
-		Tensor tensor = tensors[0];
-		for (int i = 1; i < tensors.length; i++)
-			tensor = addTwo(tensor, tensors[i]);
-		return tensor;
-	}
-
-	private static Tensor addTwo(Tensor first, Tensor second) {
-		if (first.isContra != second.isContra)
-			throw new IllegalArgumentException("Mismatched tensor variance in addition.");
-		if (first.rank != second.rank)
-			throw new IllegalArgumentException("Mismatched tensor rank in addition.");
-		else if (!first.index.equals(second.index))
-			throw new IllegalArgumentException("Mismatched tensor index name in addition.");
-		else if (first.elements.length != second.elements.length)
-			throw new IllegalArgumentException("Mismatched tensor size in addition.");
-
-		if (first.elements[0] instanceof Tensor && second.elements[0] instanceof Tensor) {
-			GeneralFunction[] elementSums = new GeneralFunction[first.elements.length];
-			for (int i = 0; i < elementSums.length; i++)
-				elementSums[i] = Tensor.addTwo((Tensor) first.elements[i], (Tensor) second.elements[i]);
-			return new Tensor(first.index, first.isContra, first.rank, elementSums);
-		} else if (!(first.elements[0] instanceof Tensor || second.elements[0] instanceof Tensor)) {
-			GeneralFunction[] elementSums = new GeneralFunction[first.elements.length];
-			for (int i = 0; i < elementSums.length; i++)
-				elementSums[i] = minimalSimplify(new Sum(first.elements[i], second.elements[i]));
-			return new Tensor(first.index, first.isContra, first.rank, elementSums);
-		} else
-			throw new IllegalArgumentException("Mismatched tensor dimensions in addition.");
-	}
 
 	public Tensor executeInternalSums() {
 		String[] oldIndices = getIndices();
@@ -365,16 +275,6 @@ public class Tensor extends GeneralFunction {
 				newVariances,
 				newElements.array
 		);
-	}
-
-	private static GeneralFunction sumArbitrary(GeneralFunction... elements) {
-		if (elements[0] instanceof Tensor) {
-			Tensor tensor = (Tensor) elements[0];
-			for (int i = 1; i < elements.length; i++)
-				tensor = addTwo(tensor, (Tensor) elements[i]);
-			return tensor;
-		} else
-			return new Sum(elements);
 	}
 
 	private static int[] copyToArraySkipping(int[] oldArray, int first, int second) {
